+++ packages/blob/src/index.node.test.ts
@@ -576,6 +576,52 @@ describe('blob client', () => {
       expect(headers['x-cache-control-max-age']).toEqual('60');
     });
 
+    it('throws when filepath is too long', async () => {
+      await expect(
+        put('a'.repeat(951), 'Test Body', {
+          access: 'public',
+        }),
+      ).rejects.toThrow(
+        new Error('Vercel Blob: pathname is too long, maximum length is 950'),
+      );
+    });
+
+    it('throws when pathname contains #', async () => {
+      await expect(
+        put('foo#bar.txt', 'Test Body', {
+          access: 'public',
+        }),
+      ).rejects.toThrow(
+        new Error(
+          'Vercel Blob: pathname cannot contain "#", please encode it if needed',
+        ),
+      );
+    });
+
+    it('throws when pathname contains ?', async () => {
+      await expect(
+        put('foo?bar.txt', 'Test Body', {
+          access: 'public',
+        }),
+      ).rejects.toThrow(
+        new Error(
+          'Vercel Blob: pathname cannot contain "?", please encode it if needed',
+        ),
+      );
+    });
+
+    it('throws when pathname contains //', async () => {
+      await expect(
+        put('foo//bar.txt', 'Test Body', {
+          access: 'public',
+        }),
+      ).rejects.toThrow(
+        new Error(
+          'Vercel Blob: pathname cannot contain "//", please encode it if needed',
+        ),
+      );
+    });
+
     const table: [string, (signal: AbortSignal) => Promise<unknown>][] = [
       [
         'put',
@@ -658,5 +704,59 @@ describe('blob client', () => {
         }).rejects.toThrow(BlobRequestAbortedError);
       },
     );
+  });
+
+  // Some folks are trying to upload plain objects which cannot work, example: https://github.com/vercel/storage/issues/637
+  describe('rejects when body is incorrect', () => {
+    type TestCase = [string, () => Promise<unknown>];
+
+    const testCases: TestCase[] = [
+      [
+        'put()',
+        () =>
+          // @ts-expect-error: Runtime check for DX
+          put('foo.txt', { file: 'value' }, { access: 'public' }),
+      ],
+      [
+        'multipart put()',
+        () =>
+          put(
+            'foo.txt',
+            // @ts-expect-error: Runtime check for DX
+            { file: 'value' },
+            { access: 'public', multipart: true },
+          ),
+      ],
+      [
+        'uploadPart()',
+        () =>
+          uploadPart(
+            'foo.txt',
+            // @ts-expect-error: Runtime check for DX
+            { file: 'value' },
+            { access: 'public', key: 'foo.txt', uploadId: '1', partNumber: 1 },
+          ),
+      ],
+    ];
+
+    it.each(testCases)('on %s', async (_, operation) => {
+      await expect(operation).rejects.toThrow(
+        new Error(
+          "Vercel Blob: Body must be a string, buffer or stream. You sent a plain JavaScript object, double check what you're trying to upload.",
+        ),
+      );
+    });
+  });
+
+  describe('copy', () => {
+    it('throws when filepath is too long', async () => {
+      await expect(
+        copy('source', 'a'.repeat(951), {
+          access: 'public',
+        }),
+      ).rejects.toThrow(
+        new Error('Vercel Blob: pathname is too long, maximum length is 950'),
+      );
+    });
   });
 });