+++ test/next/src/pages/vercel/pages/blob/image.tsx
@@ -25,7 +25,7 @@ export default function Blob(props: vercelBlob.PutBlobResult): JSX.Element {
       <h1 className="text-xl mb-4">Render an upload image on the browser</h1>
 
       {/* eslint-disable-next-line @next/next/no-img-element -- we want an image element here, fine */}
-      <img alt="test" src={props.url} />
+      <img alt="test" id="test-screenshot" src={props.url} />
     </div>
   );
 }