+++ packages/edge-config/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @vercel/edge-config
 
+## 1.4.0
+
+### Minor Changes
+
+- d7ef349: Add the `consistentRead` option to allow reading from the origin. Note that it's not recommended to use this property without good reason due to the extrem performance cost.
+
+## 1.3.0
+
+### Minor Changes
+
+- aaec8c5: Support new connection string format
+
+## 1.2.1
+
+### Patch Changes
+
+- 3057a36: gracefully handle when an empty string is supplied as the key
+
+  - `get("")` will return `undefined`
+  - `has("")` will return `false`
+  - `getAll(["a", ""])` will ignore the empty string
+
+## 1.2.0
+
+### Minor Changes
+
+- 6a592b5: allow setting fetch cache behaviour
+
+### Patch Changes
+
+- 6a592b5: remove DeepReadOnly type
+
+## 1.1.1
+
+### Patch Changes
+
+- 585a753: Resolved bug where an unhandled promise rejection event may have been triggered during development
+
 ## 1.1.0
 
 ### Minor Changes